Refactor Bot Command base class to split code into traits
